+++ lib/osm/member.rb
@@ -380,6 +380,12 @@ module Osm
       return "#{first_name}#{seperator.to_s}#{last_name}"
     end
 
+    # Check if the member is in the leaders grouping
+    # @return [Boolean]
+    def leader?
+      grouping_id.eql?(-2)
+    end
+
     # Get the Key to use in My.SCOUT links for this member
     # @param [Osm::Api] api The api to use to make the request
     # @return [String] the key