+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Version 1.2.12
 
- * 
+ * Add leader? method to member (true if member is in the leaders grouping)
 
 ## Version 1.2.11